improve divert logic if chroot was closed unexpectedly

+++ tmp/ubuntu-scratch/scripts/chroot.sh
@@ -5,15 +5,21 @@ mountfs "chroot/dev"
 mountfs "chroot/proc"
 mountfs "chroot/sys"
 
-ebegin "divert initctl (fix upstart bug)"
-	chroot chroot /usr/bin/dpkg-divert --quiet --local --rename --add /sbin/initctl
-	ln -s /bin/true chroot/sbin/initctl
-eend $?
+if [[ ! -f chroot/sbin/initctl.distrib ]];
+then
+	ebegin "divert initctl (fix upstart bug)"
+		chroot chroot /usr/bin/dpkg-divert --quiet --local --rename --add /sbin/initctl
+		ln -s /bin/true chroot/sbin/initctl
+	eend $?
+fi
 
-ebegin "divert invoke-rc.d (fix apt-get bug with PI-scripts)"
-	chroot chroot /usr/bin/dpkg-divert --quiet --local --rename --add /usr/sbin/invoke-rc.d
-	ln -s /bin/true chroot/usr/sbin/invoke-rc.d
-eend $?
+if [[ ! -f chroot/usr/sbin/invoke-rc.d.distrib ]];
+then
+	ebegin "divert invoke-rc.d (fix apt-get bug with PI-scripts)"
+		chroot chroot /usr/bin/dpkg-divert --quiet --local --rename --add /usr/sbin/invoke-rc.d
+		ln -s /bin/true chroot/usr/sbin/invoke-rc.d
+	eend $?
+fi
 
 ebegin "generate dbus machine id"
 	chroot chroot dbus-uuidgen > /var/lib/dbus/machine-id
@@ -29,15 +35,21 @@ chroot chroot /usr/bin/env -i \
   /bin/bash --login +h
 
 # clean and restore
-ebegin "restore initctl"
-	rm -f chroot/sbin/initctl
-	chroot chroot /usr/bin/dpkg-divert --quiet --rename --remove /sbin/initctl
-eend $?
+if [[ -f chroot/sbin/initctl.distrib ]];
+then
+	ebegin "restore initctl"
+		rm -f chroot/sbin/initctl
+		chroot chroot /usr/bin/dpkg-divert --quiet --rename --remove /sbin/initctl
+	eend $?
+fi
 
-ebegin "restore invoke-rc.d"
-	rm -f chroot/usr/sbin/invoke-rc.d
-	chroot chroot /usr/bin/dpkg-divert --quiet --rename --remove /usr/sbin/invoke-rc.d
-eend $?
+if [[ -f chroot/usr/sbin/invoke-rc.d.distrib ]];
+then
+	ebegin "restore invoke-rc.d"
+		rm -f chroot/usr/sbin/invoke-rc.d
+		chroot chroot /usr/bin/dpkg-divert --quiet --rename --remove /usr/sbin/invoke-rc.d
+	eend $?
+fi
 
 ebegin "clean temporary stuff"
 	find chroot/usr/share/icons -type f -name icon-theme.cache -delete
